Remove refactor to CreateConnectorWidget component
