--- frontend/packages/topology/src/components/CreateConnectorWidget.scss
@@ -1,3 +1,0 @@
-.topology-create-connector {
-  cursor: pointer;
-}
