+++ frontend/packages/topology/src/components/index.ts
@@ -5,5 +5,4 @@ export { default as ConnectorArrow } from './ConnectorArrow';
 export { default as EdgeConnectorArrow } from './EdgeConnectorArrow';
 export { default as GraphComponent } from './GraphComponent';
 export { default as VisualizationSurface } from './VisualizationSurface';
-export * from './CreateConnectorWidget';
 export { default as DefaultCreateConnector } from './DefaultCreateConnector';
